help: Use unordered lists for sub-lists in saml-authentication.md.
Our current conversion script does not work well with inserting
flattened steps in indented lists. Adding support for indented lists
would be non-trivial. Flattened steps are inserted either in case of
files or relative/setting links. I used this (  +)\{ regex to test for
instances where we might need to insert Flattened steps in an indented
block. Turns out saml-login-button.md is the only includes that faces
the issue only in saml-authentication.md.

There was no strong reason for these sub-lists to be ordered lists and
we decided to turn it into unordered list. Discussion at
https://chat.zulip.org/#narrow/channel/19-documentation/topic/new.20help.20center.3A.20numbering.20regression/near/2232242