left_sidebar: Clean up hide_unread_counts vs hide-unread-messages-count.
Now the former is only used in the channels list (headers and inactive
toggles included) and the latter is only used for the home view.

Note that this changes some behavior, notably the header and inactive
toggles now show the masked dots when all unread markers in them are
also masked.