help: Update for mobile "Copy link to..." actions.
Adds mobile instructions for copying a link to a message, topic,
and channel via the long-press action menus. Clarifies some of
the text around link formatting being in the web/desktop apps.

Fixes some example links to redirect to headers on the page, and
adds a missing period to a sentence.