generate-integration-docs-screenshot: Use bot-name option to set email.
Previously, only one bot can be created using the
generate-integration-docs-screenshot script for each integration. This
was because the bot's email was constructed using the integration's
name.

Thus, the bot-name commandline option would not work if a bot already
exists for the integration.

Now, we create the email by cleaning the bot-name value, if passed.