onboarding: Fix focused to compose recipient box when modal is closed.
When a new user visits zulip for the first time:
* they are narrowed to DM with welcomer bot
* shown a navigation_tour_video in a modal

Earlier, when the modal was closed the compose recipient box
was focused instead of the compose textarea.

Reason:
`narrow_to_dm_with_welcome_bot_new_user` triggers a focus change from
(see `compose_actions.show_compose_box` with `opts.defer_focus = true`).

We start initializing this modal while the focus transition is in
progress, resulting in a flaky behaviour of the element that will
be in focus when modal is closed. Because the modal maintains the
focus position before and after toggling modal.

This commit fixes the issue by explicitly setting the focus to
ompose textarea when the modal is closed.