compose_paste: Wrap single line content in language block.
We now wrap single liner code content in a language block if the
required metadata is present in the `paste_html`. The inline
markdown paste behavior now only applies to codeblocks that don't
have a language associated with them or miss the language
metatdata in the `paste_html`.

The latter will occur less frequently after change
that will be introduced in the following commit.

Signed-off-by: apoorvapendse <apoorvavpendse@gmail.com>