banners: Apply text-align property to banner-label element.
Adding "text-align: start" css to banner-label element makes
sure that the banner text is not aligned as per the alignment
of parent element, as we require left alignment of text for
all the banners. Only navbar banners have text align in center
but that is handled by adjusting the flex layout of the
banner-content, parent of banner-label, element.