search_suggestion: Fix empty topic suggestion topic display name.
Earlier, we removed topic display name for empty topic operand when
query was for topic autocomplete. But we should make sure that we
only do this for the last search term in the suggestion which
corresponds to the typed query and not any previously selected
entered search term. Also recently we removed subset suggestions,
so can safely check it for last search term for all suggestions.

This commit fixes this by checking if the topic suggestion is the
last search_term in the suggestion/search_terms.

Co-authored-by: Evy Kassirer <evy@zulip.com>