left_sidebar: Fix muted channel visible in collapsed folder on search.
While searching in left sidebar, if you collapse a folder, muted
or inactive channels are still visible.

This is due specificity of hiding the channel when folder is
collapsed was lower than that of showing it when a search
is active.

Fixed by increasing the specificity of hiding it when folder
is collapsed.