stream-settings: Do not remove private stream row always.
Previously, private stream row was removed from the relevant
tab view in left panel in stream settings UI when a user was
subscribed or unsubscribed. Now we only remove the row if the
user does not permission to toggle the subscription state just
like we do for public streams where we keep the row as user
can toggle the subscription state again.

Also updated the comment in update_stream_row_in_settings_tab
to be more clear about the behavior.