compose: Don't check for posting policy errors without validation.
Since `compose_validate.validate` sets the error messages
for posting policy, it is important that it is called before
we check for posting policy errors.

To fix it, we remove independent calls to
check_posting_policy_for_compose_box and call it after validation.

This also fixes a bug where compose posting policy error is
not displayed when a user is trying to DM a recipient with no
previous message history and the user is not allowed to start
a new DM conversation.

Co-authored-by: Lauryn Menard <lauryn@zulip.com>