dark_theme: Clean up feedback container styles.
The feedback container doesn't have a border so it doesn't make sense
to have a border color defined for the dark theme, and the hover color
change on the `a` tag was an unintended side effect of
commit 650edf39d0b09ff60e07d083bdcc22904a9b3315, which while correcting
the hover effect on the typeaheads, also affected the link hover styles
in the feedback container.

This commit, due the reasons mentioned above, removes the entire CSS
block related to the feedback container from web/styles/dark_theme.css.

Fixes part of #35880.