help: Move static/images/help to starlight_help/src/images.
Astro was importing these images and doing it's own processing anyways,
there was not a reason for these images to stay in the static folder.

For channels-and-topics.png, we keep the help center directory as the
source directory for two symlinks since that is where the image is
likely to stay the longest.