widgets: Convert module to typescript.
This commit fixes the return type of `zform.activate` which
was returning an object with `handle_events` instead of just
`handle_events`, which was different than the other widgets.

It also returns `handle_events` function is the extra data in
`poll_widget` isn't parsed correctly, though we expect it to
always be the right type, and log an error if it's not.

This commit changes what's sent to the server in
`make_server_callback`, since several existing type signatures
suggest we shouldn't be sending `undefined`, and it seems like
the server process_submessage doesn't expect undefined values
for `content`.