push_notification: Skip clearing notif for users without push devices.
Previously, we enqueued events to "missedmessage_mobile_notifications"
with type="remove" to clear push notificiations already sent to users.

We used to do it even for users who had no registered push devices.
'handle_remove_push_notification' later used to perform the check &
skip if there were no registered devices.

21dedfe40a02cc64a made improvements to skip enqueuing events
to send push notif if no push device is registered - it helped to
lower the number of events enqueued to *clear* notifs as well because
we have overall less number of usermessages with
'active_mobile_push_notification' flag set.

But there's still a couple of cases where events to clear can be
enqueued when no push device is registered for a user:
* After 21dedfe40a02cc64a:
  * event to send push notif is enqueued when push device was present
  * push device unregistered
  * event to clear (type="remove") can be enqueued as usermessages
    with active_mobile_push_notification flag set is present.
* event to send push notif enqueued prior to 21dedfe40a02cc64a results
  in usermessages with 'active_mobile_push_notification' flag set -
  so events to clear them can be enqueued post upgrade.

This commit adds check to skip enqueuing events with type="remove"
if no push device is registered for the user.

By doing so, we reduce unnecessary churn.

Signed-off-by: Prakhar Pratyush <prakhar@zulip.com>