help: Make `include` a sibling of `content/docs` directory.
When moving content between the include files and the docs files while
copy pasting them, there was a need to change the imports everytime
this was done. This was inconvenient to do.

We considered path aliases to solve this problem but not all editors
support path resolution via those aliases and thus we change the
directory structure instead.

We change imports for components, include files and images in this
commit.

See
https://chat.zulip.org/#narrow/channel/19-documentation/topic/confused.20about.20include.20file
for more info.