demo-orgs: Remove demo org creation from registration_helper.
With the new endpoint and view for creating demo organizations, the
only remaining use of creating demo orgs via registration_helper is
in the devlogin page.

Adds a helper function for creating demo organization realm and
owner account, which can be used to create test demos in the
devlogin page without the captcha and rate limiting checks that
are in the view function.

Removes the demo organization code from registration_helper and
the RegistrationForm.