demo-orgs: Fix configure email banner action button on modals.
Fixes the action button on the configure email banner that appears
before a demo organization owner has added an email address on the
invite user and convert demo organization modals.

Adds the banner-wrapper class so that the general CSS rules for
banners are applied.

Removes the extra padding on the button as it does not appear to
be necessary any longer as the font and line height for the banner
label and action buttons is now the same.