uploads: Move charset-detection before S3 upload, so it appears there.
We serve the content-type that we store in S3; we were only appending
the charset in the local database, and not in the S3 metadata, which
meant that we never actually provided the guessed content-type on
uploads stored in S3.  Nor does the guess-on-download codepath we have
for local storage work here, since we do not see the contents when
they are served from S3.

Append the guessed charset earlier in the process.