push_notification: Fix invalid `data` arg to firebase_messaging.Message.
Earlier, we were passing invalid `data` argument to
`firebase_messaging.Message` which would result in an error while
sending E2EE push notification for android devices.

The API requires all keys and values in the `data` dictionary
to be strings. One of the value was an integer.

This commit fixes the bug by converting the values to str if they
are not.

Signed-off-by: Prakhar Pratyush <prakhar@zulip.com>