auth: Add "next" to SOCIAL_AUTH_FIELDS_STORED_IN_SESSION.
Gets around quirky behavior of python-social-auth, which normally saves
"next" param value in the session due to it being the redirect field -
but doesn't clear it from the session when the param is absent.
By adding it to `SOCIAL_AUTH_FIELDS_STORED_IN_SESSION`, we make it
subject to the usual mechanism, which clears values from the session if
the corresponding params are not used in the current auth request.