stream-settings: Fix live updating of stream privacy options.
There was a bug where the private stream options were live
updated when a user gained access to create private streams
even if the stream was set as a default stream which cannot
be changed to private.

This was because update_private_stream_privacy_option_state
used "false" as default for "is_default_stream" parameter.
This commit updates the code to compute is_default_stream
value in update_private_stream_privacy_option_state itself
to avoid any such bugs and this also helps in refactoring
in further commits.