settings: Refactor code to update default stream and privacy UI.
This commit updates the code to live update stream private stream
options and default stream option in stream settings UI to use
separate functions.

Previously, a single function was used in most of the cases which
updated the default stream option and then called the function to
update private stream options. This is a preparatory commit
to help organize the code for disabling the private stream options
based on the other conditions which does not affect default stream
option.