recent_view: Fix incompletely updated UI after topic edits.
If the new message for topic edit arrives before the topic update
event, the recent view ui ends up in a broken state due to the
way we process these events.

Best way to avoid this bug and similar race bugs of this kind is
just to do a complete rerender.