invite: Fix user invite user modal tab toggle hidden bug.
The invite user modal's tab toggler was being hidden when all the steps
mentioned in the invite tips banner were fulfilled. This happened due
to an error in logic in commit ed304c8af418fa022c6f88bc85087207328185b4
which returned early in the common codepath, when the banner was not to
be shown, preventing the tab toggler from being initialized & rendered.