compose: Remove unused CSS for compose textarea.
We do not disable the compose textarea or make it readonly
so this commit removes the CSS for setting background color
of disabled compose textarea.

Since there is no border for compose textarea there is no
need for border-color CSS set for focussed compose textarea
in dark theme and this commit removes that as well.