middleware: If there are no realms yet, show proxy errors to end-user.
The most common case for configuring Zulip behind a proxy is that
there are no realms on the server -- and attempting to load anything
causes a frustrating 500 server error page.

In cases where there are no realms (including the system realm, since
that is only made when the first realm realm is), proxy
misconfigurations result in showing the end-user debugging information
and suggested next steps.  While there is potentially a very small
chance that doing so exposes internal network configuration details to
an potentially malicious external end-user, the window of doing so is
very limited, the impact is low, and the benefits to the administrator
are high.