message_view: Fix selected message is changed on server reload.
Fixes #36577

Reproducer

(1) Highlight a message that isn't the last one in the narrow
(2) Change code in the backend (e.g. add x = 3 to a file)
(3) See the last message is now highlighted.
    Expectation would be that the middle message is highlighted.

This was happening since `then_select_id` was being ingored since
we didn't know if user is in the correct narrow and hence focused
on just fetch the narrow messages using the `with` term.
To fix it, we first fetch the `with` term message, adjust the
narrow if required and then show the message feed.