i18n: Write language name in the language itself.
There were a couple of languages whose name were written in
English in the language dropdown and modal. This commit updates
code to write their name in the language itself using
"Intl.DisplayNames".

We also add region and script to show in brackets after the
language so "British English" is written as "English (United Kingdom)"
and the scripts are added in brackets in Chinese languages.
Above two are just examples, there are a couple of more langauges
where region/script is added to the language name.

Fixes #14565.