settings: Disable realm setting if disabled in server.
We disable the checkboxes for settings responsible for
allowing non-admins to change their name or avatar if
the server level setting does not allow changing them.

We also keep the checkbox checked to make it clear to
users that they cannot change their name or avatar
as it is not allowed by the server setting.

Fixes #12286.