settings_profile_fields: Rename functions for readability.
To distinguish between the similarly named functions,
- set_up_form_fields -> initialize_custom_profile_field_form as
this function is called only once upon form initialization.
- set_up_create_field_form -> update_form_for_field_type_selection as
this function is called every time the field type is changed.

Renamed the 2 modal manager to use a common prefix and mention the form
names that they correspond to - "create" and "edit".
- open_edit_form_modal -> open_custom_profile_field_edit_form_modal
- open_custom_profile_field_form_modal ->
open_custom_profile_field_creation_form_modal

- clear_form_data -> initialize_form_to_defaults as the function is
used for initalization and re-initialization.