stream-settings-data: Fix comment for get_streams_for_settings_page.
The TODO comment was no longer valid since we no longer use this
function for copy-from-stream feature and instead this function
is used to get options for the notification settings dropdowns.

Also, fixed another comment to mention about unsubscribed streams
as well along with subscribed streams since this function returns
both of them.