css: Fix min-height of textarea elements.
We need to set min-height of textarea elements such that it
is enough for two lines of text. It was set to 2em which
did not consider line-height value and box-sizing values.

So, this commit fixes the code to set the min-height to
"2em * line-height" for elements with box-sizing set to
"content-box" and include paddings and borders for elements
with box-sizing set to "border-box".