message_scroll: Restyle scroll-to-bottom button.
This commit updates the scroll-to-bottom button to use the new action
button component, with tweaks to the border radius to achieve a rounded
button.

A backdrop blur is applied on hover to improve visibility against
varying message backgrounds, and on top of the message timestamps.

The clip-path property is also removed in the active state since we
don't desire the indented effect when the button is clicked.

Fixes #36640.