stream-list-sort: Simplify topic check for current narrow channel.
Use new helper function, topic_list_data.get_filtered_topic_names,
to bulid a list of topics in the current narrow channel that match
the search term. This list can then be checked for unmuted topics
for whether to include the channel in the list of matches.

Previously, we were building the topic list for the filtered by
search term topics to just check to see that any topic items had
been returned.