settings: Duplicate bots panel in "Personal" tab.
Previously, clicking on "Bots" in "Personal" tab redirected
to bots panel in "Organization" tab. This commit changes this
behavior to not redirect and instead show the same content
shown in bots panel in "Organization" tab in the "Personal"
tab and opening "Your bots" section by default instead of
"All bots".

Now "Bots" in both "Personal" and "Organization" opens
different sections independent of each other maintaining
their individual filter and search value selected.

This helps in maintaining the keyboard navigation using
up/down keys in the settings sidebar which was broken
due to redirecting.

We also now show lock icon for the personal bots panel
if user does not have permission to create or manage
bots like we do for organization bots panel.