user_profile: Handle & update show_user_group_container from groups tab.
This now renders the widget for adding groups always, instead of only when
`show_user_group_widget` is true. This doesn't cause security issues because
`user_group_picker_pill.set_up_pill_typeahead` checks for groups with
correct permissions before populating the widget.

Having the widget always in the DOM makes it easier to show/hide it when
there are/aren't groups that can be added. Before this commit, the widget
would always show even if there were no groups that could be added at the
time.