narrow: Reduce find_first_unread_anchor query count.
Previously, find_first_unread_anchor used to call
get_base_query_for_search unconditionally since we wanted to set
need_user_message to True and assumed that since we are just fetching
the query, there should be no extra calls to doing that.

But if need_user_message is True, we call
get_recursive_membership_groups, which is one query. If
need_user_message was already set to True, this is a wasteful DB call.
So this commit changes find_first_unread_anchor to only rebuild the
query in case of need_user_message being false.

The ideal thing would be to just have the CTE in the query being
returned by get_base_query_for_search. But
get_recursive_membership_groups uses a Django CTE that cannot be
reused in get_base_query_for_search since it uses SQLAlchemy. We can
replicate the logic being done by get_recursive_membership_groups for
our case, but it is not worth it to maintain another copy of
get_recursive_membership_groups just to save 1 query count.