composebox_typeahead: Change topic typeahead to always open upwards.
Currently, in topic recipient input in compose box, if the topic
typeahead ahead is small enough then the typeahead is placed below the
input field. And when the typeahead list has more entries then it would
would be placed above the input field. This typeahead shifts places
whenever user types in the input.

We fix the typeahead for topic recipients to always opened upwards.