settings: Allow tabs and buttons in right panel to wrap.
Previously, the tabs and buttons in the right panel for a
selected stream or group would wrap into next line only for
smaller screen width. This resulted in all tabs and buttons
showed in a single line with a scrollable view for larger
font sizes which did not look good. This commit updates the
CSS to allow wrapping to next line for larger screen width
so that the buttons can move to next line on larger font
size.