digest_email: Fix error on including archived channel messages.
Earlier, we were experiencing an assertion error when messages
from archived channels were to be included in digest emails.

`assert stream_id in stream_id_map` used to fail in
`digest_block_header`.

This was due to `stream_id_map` only including `active` channels
of the realm.

This commit fixes the bug by updating `stream_id_map` to include
channels present in `user_stream_map`, irrespective of whether
they are archived or not.

There's no need to fetch all channels in the realm.

Signed-off-by: Prakhar Pratyush <prakhar@zulip.com>