gifs: Unify GIPHY and Tenor UI.
This is a big bang commit that combines the
GIPHY picker UI with the custom grid implementation
and completely removes the existing third party
component.

We introduce the GiphyNetwork class as a part
of this commit.

We, also rename some of the CSS classes that would
have otherwise broken the behavior of the GIPHY
picker if done in previous commits.

We rename `tenor.ts` to `gif_picker_ui.ts` and
introduce a `provider` state variable for some
conditional logic in `toggle_picker_popover`.

Signed-off-by: apoorvapendse <apoorvavpendse@gmail.com>