flatpickr: Fix tab navigation when time is disabled.
Elements that were only visible when time was enabled were part of the
tab navigation list without any condition. Now we add them only when
time is enabled.

We also changed how we set the default for enableTime. Before this
commit we were relying on options to override the default `enableTime:
true`, which means that we cannot use the actual enableTime value in
any other arg when initializing the flatpickr instance. This is in
preparation of adding the `Jump to date` button.